make bu_ptbl_trunc take a size_t instead of an int for better intent-based comparison and warning-free compiles
svn:revision:67855
svn:branch:trunk
svn:account:brlcad

+++ src/libbu/ptbl.c
@@ -264,7 +264,7 @@ bu_pr_ptbl(const char *title, const struct bu_ptbl *tbl, int verbose)
 
 
 void
-bu_ptbl_trunc(struct bu_ptbl *tbl, int end)
+bu_ptbl_trunc(struct bu_ptbl *tbl, size_t end)
 {
     BU_CK_PTBL(tbl);
